+++ src/ua/data_types/browse_next_response.rs
@@ -1,0 +1,19 @@
+use crate::{ua, ServiceResponse};
+
+crate::data_type!(BrowseNextResponse);
+
+impl BrowseNextResponse {
+    #[must_use]
+    pub fn results(&self) -> Option<ua::Array<ua::BrowseResult>> {
+        // TODO: Adjust signature to return non-owned value instead.
+        ua::Array::from_raw_parts(self.0.results, self.0.resultsSize)
+    }
+}
+
+impl ServiceResponse for BrowseNextResponse {
+    type Request = ua::BrowseNextRequest;
+
+    fn service_result(&self) -> ua::StatusCode {
+        ua::StatusCode::new(self.0.responseHeader.serviceResult)
+    }
+}